+++ feat/my/src/main/java/com/emotionstorage/my/ui/component/dialog/WhenToUseKeyDialog.kt
@@ -1,4 +1,4 @@
-package com.emotionstorage.my.ui.component
+package com.emotionstorage.my.ui.component.dialog
 
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
